Deleted unused classes and added logic to bll

+++ LightMessanger.BLL/Services/GroupsService.cs
@@ -16,7 +16,9 @@ namespace LightMessanger.BLL.Services
         {
             if (item.Name.Length < 3 || item.Name.Length > 30)
                 throw new ArgumentException("Invalid name");
-            await _context.AddAsync(item);
+            if (await GetValueByСonditionAsync(u => u.Name, item.Name) != null)
+                throw new ArgumentException("GroupName exist");
+                await _context.AddAsync(item);
         }
         public async Task<IEnumerable<Group>> SearchBySubstringInNameAsync(string value)
         {
@@ -27,15 +29,23 @@ namespace LightMessanger.BLL.Services
             return await _context.GetValueByСonditionAsync(valueSelector, value);
         }
 
-        public async Task AddUserInGroup(User user, int IdGroup)
+        public async Task AddUserInGroup(User user, string groupName)
         {
-            var group = await _context.GetByIdAsync(IdGroup);
+            var group = await GetGroupWithUsers(groupName);
+            if (group is null)
+                throw new ArgumentNullException("Group not found");
+            if (group.Users.Any(x => x.Id.Equals(user.Id)))
+                throw new ArgumentException("User already exist in group");
+          
+
             group.Users.Add(user);
             await _context.UpdateAsync(group);
         }
 
         public async Task DeleteAsync(Group item)
         {
+            if (item is null)
+                throw new ArgumentNullException("Group not exist");
             await _context.DeleteAsync(item);
         }
 
@@ -51,12 +61,15 @@ namespace LightMessanger.BLL.Services
 
         public async Task UpdateAsync(Group item)
         {
+            if (item.Name.Length < 3 || item.Name.Length > 30)
+                throw new ArgumentException("Invalid name");
+           
             await _context.UpdateAsync(item);
         }
 
         public async Task<Group> GetGroupWithUsers(string name)
         {
-            return await _context.GetValueByСonditionAsync(u=>u.Name, name,new List<string>() { 
+            return await _context.GetValueByСonditionAsync(u => u.Name, name, new List<string>() {
             "Users",
             "GroupMessages"
             });
